refactor: update imports and remove unused functions
- Changed import path from "../../src/data-files" to "../../src/datafile" in sections.test.ts.
- Marked "hasSections" and "parseSections" tests as todo in sections.test.ts.
- Removed "hasSections" and "parseSections" from exports snapshot in exports.test.ts.
- Updated import path for datafile in exports snapshot.
- Removed "./src/data-files.ts" from tsdown.config.ts entry.
