+++ src/datafile.ts
@@ -1,2 +1,4 @@
 export * from "./datafile/ast";
+export { RawDataFile } from "./datafile/model";
+export { hasSections, parseSections } from "./datafile/sections";
 export * from "./datafile/typeguards";
