+++ test/exports.test.ts
@@ -14,7 +14,6 @@ it("exports-snapshot", async () => {
       fromHexToCodepoint: "function",
       getBoundaryLineStyle: "function",
       getCurrentDraftVersion: "function",
-      hasSections: "function",
       hasUCDFolderPath: "function",
       inferFileName: "function",
       inferVersion: "function",
@@ -27,10 +26,8 @@ it("exports-snapshot", async () => {
       isHashBoundary: "function",
       isLineWithData: "function",
       isMissingAnnotation: "function",
-      parseDataFileHeading: "function",
       parseFileNameLine: "function",
       parseMissingAnnotation: "function",
-      parseSections: "function",
       RawDataFile: "function",
       resolveUCDVersion: "function",
       stripHex: "function",
@@ -46,13 +43,9 @@ it("exports-snapshot", async () => {
       UNICODE_STABLE_VERSION: "string",
       UNICODE_VERSION_METADATA: "object",
     },
-    "./data-files": {
-      hasSections: "function",
-      parseDataFileHeading: "function",
-      parseSections: "function",
-      RawDataFile: "function",
-    },
     "./datafile": {
+      hasSections: "function",
+      RawDataFile: "function",
       NodeTypes: "object",
       isBoundaryNode: "function",
       isCommentNode: "function",
@@ -61,6 +54,7 @@ it("exports-snapshot", async () => {
       isNode: "function",
       isRootNode: "function",
       isUnknownNode: "function",
+      parseSections: "function",
     },
     "./line-helpers": {
       getBoundaryLineStyle: "function",
